+++ GrailDiary/MarkdownEditingTextView.swift
@@ -4,17 +4,9 @@ import Logging
 import MobileCoreServices
 import os
 import UIKit
+import UniformTypeIdentifiers
 
 private let log = OSLog(subsystem: "org.brians-brain.ScrapPaper", category: "TextView")
-
-public protocol MarkdownEditingTextViewImageStoring: AnyObject {
-  /// The text view has an image to store because of a paste or drop operation.
-  /// - parameter textView: The text view
-  /// - parameter imageData: The image data to store
-  /// - parameter suffix: Image data suffix that identifies the data format (e.g., "jpeg", "png")
-  /// - returns: A string key that can locate this image later.
-  func markdownEditingTextView(_ textView: MarkdownEditingTextView, store imageData: Data, suffix: String) throws -> String
-}
 
 /// Custom UITextView subclass that overrides "copy" to copy Markdown.
 // TODO: Move renderers, MiniMarkdown text storage management, etc. to this class.
@@ -44,21 +36,21 @@ public final class MarkdownEditingTextView: UITextView {
     super.paste(itemProviders: itemProviders)
   }
 
-  public weak var imageStorage: MarkdownEditingTextViewImageStoring?
+  public var imageStorage: ImageStorage?
 
   override public func paste(_ sender: Any?) {
     if let image = UIPasteboard.general.image, let imageStorage = self.imageStorage {
       Logger.shared.info("Pasting an image")
       let imageKey: String?
-      if let jpegData = UIPasteboard.general.data(forPasteboardType: kUTTypeJPEG as String) {
+      if let jpegData = UIPasteboard.general.data(forPasteboardType: UTType.jpeg.identifier) {
         Logger.shared.info("Got JPEG data = \(jpegData.count) bytes")
-        imageKey = try? imageStorage.markdownEditingTextView(self, store: jpegData, suffix: "jpeg")
-      } else if let pngData = UIPasteboard.general.data(forPasteboardType: kUTTypePNG as String) {
+        imageKey = try? imageStorage.storeImageData(jpegData, type: .jpeg)
+      } else if let pngData = UIPasteboard.general.data(forPasteboardType: UTType.png.identifier) {
         Logger.shared.info("Got PNG data = \(pngData.count) bytes")
-        imageKey = try? imageStorage.markdownEditingTextView(self, store: pngData, suffix: "png")
+        imageKey = try? imageStorage.storeImageData(pngData, type: .png)
       } else if let convertedData = image.jpegData(compressionQuality: 0.8) {
         Logger.shared.info("Did JPEG conversion ourselves = \(convertedData.count) bytes")
-        imageKey = try? imageStorage.markdownEditingTextView(self, store: convertedData, suffix: "jpeg")
+        imageKey = try? imageStorage.storeImageData(convertedData, type: .jpeg)
       } else {
         Logger.shared.error("Could not get image data")
         imageKey = nil