+++ GrailDiary/Sqlite/NoteMetadataRecord.swift
@@ -12,7 +12,20 @@ public struct NoteMetadataRecord: Decodable, FetchableRecord {
   var folder: String?
   var noteLinks: [NoteLinkRecord]
   var contents: [ContentRecord]
+  var thumbnailImage: [BinaryContentRecord]
   var summary: String?
+  var image: String?
 
-  static let request = NoteRecord.including(all: NoteRecord.noteHashtags)
+  /// Includes all of the joins you need to get the base information for a NoteMetadataRecord.
+  /// You can chain other filters on this.
+  static func request(baseRequest: QueryInterfaceRequest<NoteRecord> = NoteRecord.all()) -> QueryInterfaceRequest<NoteMetadataRecord> {
+    let referenceRecords = NoteRecord.contentRecords.filter(ContentRecord.Columns.role == ContentRole.reference.rawValue)
+    let thumbnailImages = NoteRecord.binaryContentRecords.filter(BinaryContentRecord.Columns.role == ContentRole.embeddedImage.rawValue).forKey("thumbnailImage")
+    return baseRequest
+      .filter(NoteRecord.Columns.deleted == false)
+      .including(all: NoteRecord.noteHashtags)
+      .including(all: referenceRecords)
+      .including(all: thumbnailImages)
+      .asRequest(of: NoteMetadataRecord.self)
+  }
 }